+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Druids and Fighters can now spend hit dice during a breather to recover Wild Shape and Second Wind uses
+- Spellcasters can now spend hit dice during a breather to recover spell slots (1 HD per spell level)
+
 ## [13.1.0](https://github.com/nivthefox/foundryvtt-sosly-5e-house-rules/releases/tag/13.1.0) - 2026-01-11
 
 ### Added